Fixed style nits.
Style fixes to reflect belated review feedback on http://codereview.chromium.org/340036.

BUG=26233
TEST=None

Review URL: http://codereview.chromium.org/342064

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@30630 0039d316-1c4b-4281-b951-d872f2087c98

+++ webkit/api/public/WebSharedWorkerRepository.h
@@ -33,7 +33,6 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #define WebSharedWorkerRepository_h
 
 #include "WebCommon.h"
-#include "WebVector.h"
 
 namespace WebKit {
     class WebString;
@@ -43,7 +42,7 @@ namespace WebKit {
     class WebSharedWorkerRepository {
     public:
         // Unique identifier for the parent document of a worker (unique within a given process).
-        typedef uintptr_t DocumentID;
+        typedef unsigned long long DocumentID;
 
         // Connects the passed SharedWorker object with the specified worker thread.
         // Caller is responsible for freeing the returned object. Returns null if a SharedWorker with that name already exists but with a different URL.
